+++ src/App.js
@@ -2,25 +2,15 @@ import React from 'react';
 import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom';
 import Home from './pages/Home';
 import FoodDetailPage from './pages/FoodDetailPage';
-import LoginRegister from './components/LoginRegister/LoginRegister';
-import Survey from './components/Survey/Survey';
+import LoginRegister from './pages/LoginRegister';
+import Survey from './pages/Survey';
 
 const App = () => {
   return (
     <Router>
       <Routes>
-        {/* Default Route: "/" points to Home */}
-       {/* {<Route path="/" element={<Home />} />} */}
-
-        {/* Dynamic Route for Food Details */}
-       {/* <Route path="/food/:id" element={<FoodDetailPage />} /> */}
-
-        {/* Fallback Route: Redirect to Home if no match */}
-        {/* <Route path="*" element={<Navigate to="/" />} /> */}
-
-        <Route path="/" element={<LoginRegister />} /> {/* LoginRegister route */}
-      
-        <Route path="/Survey" element={<Survey />} /> {/* Survey route */}
+        <Route path="/" element={<Home />} /> {/* Default route */}
+        <Route path="/login" element={<LoginRegister />} /> {/* LoginRegister route */}
       </Routes>
     </Router>
   );
