--- src/components/LoginForm/LoginForm.css
@@ -1,134 +1,0 @@
-/* Wrapper styling for the login form */
-.wrapper {
-    width: 420px;
-    background: transparent;
-    border: 2px solid rgba(255, 255, 255, 0.2);
-    color: #fff;
-    border-radius: 20px;
-    padding: 30px 40px;
-    box-shadow: 0px 4px 12px rgba(0, 0, 0, 0.1);
-}
-
-/* Form title styling */
-.wrapper h1 {
-    font-size: 36px;
-    text-align: center;
-    margin-bottom: 20px;
-}
-
-/* Input box styling */
-.wrapper .input-box {
-    position: relative;
-    width: 100%;
-    height: 50px;
-    margin: 30px 0;
-}
-
-.input-box input {
-    width: 100%;
-    height: 100%;
-    background: transparent;
-    border: 2px solid rgba(255, 255, 255, 0.2);
-    outline: none;
-    border-radius: 40px;
-    font-size: 20px;
-    color: #fff;
-    padding: 0 45px 0 20px;
-}
-
-.input-box input::placeholder {
-    color: #ffffffb3; /* Slightly lighter white for placeholder text */
-}
-
-.input-box .icon {
-    position: absolute;
-    right: 20px;
-    top: 50%;
-    transform: translateY(-50%);
-    font-size: 20px;
-    color: #ffffffb3;
-}
-
-/* Remember and forgot password section */
-.wrapper .remember-forgot {
-    display: flex;
-    justify-content: space-between;
-    font-size: 16px;
-    margin: -10px 0 20px;
-}
-
-.remember-forgot label input {
-    accent-color: #fff;
-    margin-right: 5px;
-}
-
-.remember-forgot a {
-    color: #ffffffb3;
-    text-decoration: none;
-    font-size: 16px;
-}
-
-.remember-forgot a:hover {
-    text-decoration: underline;
-}
-
-/* Button styling */
-.wrapper button {
-    width: 100%;
-    background: #fff;
-    border: none;
-    outline: none;
-    border-radius: 40px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-    cursor: pointer;
-    font-size: 18px;
-    color: #333;
-    font-weight: 700;
-    padding: 12px;
-    transition: background-color 0.3s ease;
-}
-
-.wrapper button:hover {
-    background-color: #f0f0f0;
-}
-
-/* Register link styling */
-.wrapper .register-link {
-    font-size: 18px;
-    text-align: center;
-    margin: 20px 0 15px;
-}
-
-.register-link p a {
-    color: #ffffffb3;
-    text-decoration: none;
-    font-weight: 600;
-}
-
-.register-link p a:hover {
-    text-decoration: underline;
-}
-
-/* Language Switcher Styling */
-.language-switch {
-    display: flex;
-    justify-content: center;
-    gap: 10px;
-    margin-bottom: 15px;
-}
-
-.language-switch button {
-    background: transparent;
-    border: 1px solid #ffffffb3;
-    color: #ffffffb3;
-    border-radius: 20px;
-    padding: 5px 15px;
-    cursor: pointer;
-    font-size: 14px;
-    transition: color 0.3s ease, background 0.3s ease;
-}
-
-.language-switch button:hover {
-    background: #ffffffb3;
-    color: #333;
-}
