--- src/components/Survey/json.js
@@ -1,213 +1,0 @@
-export const json = {
-    "logoPosition": "right",
-    "pages": [
-      {
-        "name": "page1",
-        "elements": [
-          {
-            "type": "checkbox",
-            "name": "question2",
-            "title": "Favori mutfak türlerinizi seçiniz:\n",
-            "choices": [
-              {
-                "value": "Item 5",
-                "text": "Türk"
-              },
-              {
-                "value": "Item 1",
-                "text": "İtalyan"
-              },
-              {
-                "value": "Item 3",
-                "text": "Akdeniz"
-              },
-              {
-                "value": "Item 4",
-                "text": "Çin"
-              },
-              {
-                "value": "Item 6",
-                "text": "Hint"
-              },
-              {
-                "value": "Item 2",
-                "text": "Meksika"
-              },
-              {
-                "value": "Item 7",
-                "text": "Ortadoğu"
-              },
-              {
-                "value": "Item 8",
-                "text": "Amerikan"
-              },
-              {
-                "value": "Item 9",
-                "text": "Japon"
-              },
-              {
-                "value": "Item 10",
-                "text": "Fransız"
-              }
-            ],
-            "showSelectAllItem": true,
-            "selectAllText": "Hepsi"
-          },
-          {
-            "type": "checkbox",
-            "name": "question1",
-            "title": "Belirli bir diyet uyguluyor musunuz? Tüm uygun seçenekleri işaretleyin:\n",
-            "choices": [
-              {
-                "value": "Item 1",
-                "text": "Vejeteryan"
-              },
-              {
-                "value": "Item 2",
-                "text": "Vegan"
-              },
-              {
-                "value": "Item 3",
-                "text": "Glütensiz"
-              },
-              {
-                "value": "Item 4",
-                "text": "Laktozsuz"
-              },
-              {
-                "value": "Item 5",
-                "text": "Glütensiz2"
-              },
-              {
-                "value": "Item 6",
-                "text": "Glütensiz3"
-              },
-              {
-                "value": "Item 7",
-                "text": "Glütensiz4"
-              }
-            ],
-            "showOtherItem": true,
-            "showNoneItem": true,
-            "noneText": "Hiçbiri"
-          },
-          {
-            "type": "checkbox",
-            "name": "question3",
-            "title": "Gıda alerjiniz var mı? Uygun olanları seçiniz:\n",
-            "choices": [
-              {
-                "value": "Item 1",
-                "text": "Yer fıstığı"
-              },
-              {
-                "value": "Item 2",
-                "text": "Çerez (ör. badem, ceviz)"
-              },
-              {
-                "value": "Item 3",
-                "text": "Kabuklu deniz ürünleri"
-              },
-              {
-                "value": "Item 4",
-                "text": "Balık ve balık ürünleri"
-              },
-              {
-                "value": "Item 5",
-                "text": "Süt ürünleri"
-              },
-              {
-                "value": "Item 6",
-                "text": "Yumurta"
-              },
-              {
-                "value": "Item 7",
-                "text": "Glüten"
-              },
-              {
-                "value": "Item 8",
-                "text": "Soya"
-              },
-              {
-                "value": "Item 9",
-                "text": "Buğday"
-              },
-              {
-                "value": "Item 10",
-                "text": "Susam"
-              }
-            ],
-            "showOtherItem": true
-          },
-          {
-            "type": "matrix",
-            "name": "question4",
-            "title": "Farklı yemek türleri için tercihinizi belirtiniz:",
-            "columns": [
-              {
-                "value": "Column 1",
-                "text": "Sevmiyorum"
-              },
-              {
-                "value": "Column 2",
-                "text": "Seviyorum"
-              },
-              {
-                "value": "Column 3",
-                "text": "Çok seviyorum"
-              }
-            ],
-            "rows": [
-              {
-                "value": "Row 1",
-                "text": "Sebze Yemekleri"
-              },
-              {
-                "value": "Row 2",
-                "text": "Et Yemekleri"
-              },
-              {
-                "value": "Row 3",
-                "text": "Deniz Ürünleri"
-              },
-              {
-                "value": "Row 4",
-                "text": "Hamur İşleri ve Ekmek"
-              },
-              {
-                "value": "Row 5",
-                "text": "Tatlılar ve Şekerlemeler"
-              }
-            ]
-          },
-          {
-            "type": "checkbox",
-            "name": "question5",
-            "title": "Favori malzemeleri seçiniz:",
-            "choices": [
-              "Tavuk",
-              "Balık",
-              "Kırmızı et",
-              "Makarna",
-              "Pirinç",
-              "Patates",
-              "Peynir",
-              "Fasulye",
-              "Yeşil Zeytin ",
-              "Siyah zeytin",
-              "Soğan",
-              "Sarımsak",
-              "Taze otlar",
-              "Kuruyemiş",
-              "Meyve"
-            ]
-          },
-          {
-            "type": "text",
-            "name": "question6",
-            "title": "Bir önceki sorudaki şıklardan sevmediklerinizi yazınız"
-          }
-        ]
-      }
-    ]
-  }
