+++ src/index.css
@@ -1,19 +1,3 @@
-/*@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;600;700;&display=swap');
-
-body {
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  background: black;
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
-}*/
 @import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;600;700;&display=swap');
 
 *{
@@ -30,10 +14,16 @@ body {
   justify-content: center;
   align-items: center;
   min-height: 100vh;
-  background:white/*url("./components/Assets/background.jpg")*/
+  background:white
   no-repeat;
   background-size: cover;
   background-position: center;
+  height: 100%;
+  width: 100%; 
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+
 }
 
 code {
