+++ src/styles/LoginRegister.css
@@ -1,0 +1,164 @@
+body {
+    font-family: "Poppins", sans-serif;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    min-height: 100vh;
+
+    /* Set background to the image */
+    background: url('../Assets/login_background.jpg') no-repeat center center;
+
+    /* Ensure the image covers the entire viewport */
+    background-size: cover;
+}
+
+  
+  .wrapper {
+    position: relative;
+    background: rgba(255, 255, 255, 0.1);
+    border: 2px solid rgba(255, 255, 255, 0.2);
+    color: #fff;
+    backdrop-filter: blur(20px);
+    border-radius: 20px;
+    padding: 20px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    max-width: 450px;
+    width: 100%;
+    overflow: hidden;
+  }
+  
+  .wrapper .form-box {
+    width: 100%;
+    padding: 20px 10px;
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+  }
+  
+  .wrapper .language-switch {
+    display: flex;
+    justify-content: flex-end;
+    gap: 10px;
+    position: absolute;
+    top: 10px;
+    right: 10px;
+  }
+  
+  .wrapper  .language-switch button {
+    background: transparent;
+    border: 1px solid rgba(255, 255, 255, 0.3);
+    color: #ffffffdc;
+    border-radius: 50px;
+    cursor: pointer;
+    font-size: 13px;
+    padding: 5px 10px;
+    transition: background-color 0.3s ease, color 0.3s ease;
+  }
+  
+  .wrapper  .language-switch button:hover {
+    background: #ffffff80;
+
+  }
+  
+  .register-link p  {
+    position: relative;
+    display: flex;
+    justify-content: flex-end;
+    gap: 40px;
+    padding: 0px 60px 20px 0px;
+    text-align: center;
+  }
+
+  .register-link p span {
+    color: #ffff;
+    text-decoration: none;
+    font-weight: 600;
+
+  }
+
+  .register-link p span:hover{
+    text-decoration: underline;
+  } 
+  
+  .register-link p a {
+    font-weight: 600;
+    text-align: center;
+    padding: 0px 50px 10px 0px;
+
+  }
+
+  form h1 {
+    font-size: 28px;
+    text-align: center;
+    margin-bottom: 10px;
+    color: #ffffff;
+  }
+  
+  .input-box {
+    position: relative;
+    width: 100%;
+    height: 50px;
+    margin: 30px 0;
+  }
+  
+  .input-box input {
+    width: 100%;
+    height: 100%;
+    background: transparent;
+    border: 2px solid rgb(255, 255, 255,.2) ;
+    outline: none;
+    border-radius: 40px ;
+    font-size: 20px;
+    color: #ffff;
+    padding: 20px 45px 20px 20px;
+  }
+  
+  .input-box input::placeholder {
+    color: rgba(255, 255, 255, 0.7);
+  }
+  
+  .input-box .icon {
+    position: absolute;
+    right: 15px;
+    top: 50%;
+    transform: translateY(-50%);
+    font-size: 20px;
+    color: rgba(255, 255, 255, 0.6);
+  }
+  
+  form button {
+    width: 100%;
+    height: 50px;
+    background: #ffffff;
+    border: none;
+    border-radius: 10px;
+    font-size: 18px;
+    font-weight: bold;
+    color: #333;
+    cursor: pointer;
+    transition: background 0.3s ease;
+  }
+  
+  form button:hover {
+    background: #f0f0f0;
+  }
+  
+  .register-link {
+    text-align: center;
+    font-size: 14px;
+  }
+  
+  .register-link a {
+    color: #ffffff;
+    font-weight: bold;
+    text-decoration: none;
+  }
+  
+  .register-link a:hover {
+    text-decoration: underline;
+  }
+
